% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{picotsize_theme}
\alias{picotsize_theme}
\title{The PICOTsize bslib theme}
\format{
An object of class \code{bs_theme_with_preset} (inherits from \code{bs_version_5}, \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
picotsize_theme
}
\description{
The PICOTsize bslib theme
}
\keyword{internal}
