% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the PICOTsize Shiny app}
\usage{
run_app()
}
\value{
No return value. Launches the Shiny app in a browser or
  the RStudio Viewer pane.
}
\description{
Launch the PICOTsize Shiny app
}
\examples{
\dontrun{
run_app()
}

}
