% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Main Shiny server function for PICOTsize}
\usage{
server(input, output, session)
}
\arguments{
\item{input, output, session}{Standard Shiny server arguments.}
}
\description{
Main Shiny server function for PICOTsize
}
\keyword{internal}
